Add functionality to truncate and move large files
This commit adds a capability to limit the max file size for placing job output in elasticsearch docs. When a file is greater than the size requested, which defaults to .5MB, the end of the file is added to the doc. The full output is copied to a specified volume with the job title and timestamp as the title.