Add support for queue time
This commit adds support for getting the time a job was queued from a timestamp placed in the file. Libzpr is updated to support passing a timestamp in seconds to the get_timestamp function and return formatted time taking into account the time zone. In order to accommodate this change it is necessary to be more specific about how the title is gathered, so a function is added to handle that logic.